moved isWeakKanbanJson to storage file to circumvent "self not defined" error. This is hacky.
