+++ src/util/kanban-types.ts
@@ -154,95 +154,3 @@ export function createKanbanJson(
         timestamp: Date.now(),
     };
 }
-
-/**
- * @ignore
- */
-function matchingKeys(
-    obj: any,
-    requiredKeys: Readonly<string[]>,
-    optionalKeys: Readonly<string[]>
-): boolean {
-    if (typeof obj !== 'object') {
-        return false;
-    }
-
-    let keys = Object.keys(obj).filter((key) => !optionalKeys.includes(key));
-    for (const key of keys) {
-        if (!requiredKeys.includes(key)) {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-/**
- * @ignore
- */
-function isWeakTaskJson(obj: any): boolean {
-    if (typeof obj === 'string') {
-        return true;
-    }
-
-    if (typeof obj !== 'object') {
-        return false;
-    }
-
-    return typeof obj.id === 'string' && typeof obj.text === 'string';
-}
-
-/**
- * @ignore
- */
-function isWeakColumnJson(obj: any): boolean {
-    const mandatoryColKeys = ['title', 'tasks'] as const;
-    const optionalColKeys = ['ntasks', 'id', 'color'] as const;
-
-    if (!matchingKeys(obj, mandatoryColKeys, optionalColKeys)) {
-        return false;
-    }
-
-    if (typeof obj.title !== 'string' || !Array.isArray(obj.tasks)) {
-        return false;
-    }
-
-    for (const task of obj.tasks) {
-        if (!isWeakTaskJson(task)) {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-/**
- * Checks if `obj` is a WeakKanbanJson (or KanbanJson)
- * @param {any} obj Object to be verified
- */
-export function isWeakKanbanJson(obj: any): boolean {
-    const mandatoryKanbanKeys = ['cols'] as const;
-    const optionalKanbanKeys = [
-        'title',
-        'ncols',
-        'settings',
-        'autosave',
-        'saveToFile',
-        'timestamp',
-    ] as const;
-    if (!matchingKeys(obj, mandatoryKanbanKeys, optionalKanbanKeys)) {
-        return false;
-    }
-
-    if (!Array.isArray(obj.cols)) {
-        return false;
-    }
-
-    for (const col of obj.cols) {
-        if (!isWeakColumnJson(col)) {
-            return false;
-        }
-    }
-
-    return true;
-}
